gestion des redirections après validation des formulaires

+++ sitezinzine/src/Controller/Admin/EmissionController.php
@@ -146,12 +146,24 @@ class EmissionController extends AbstractController
 
 
     #[Route('/{id}', name: 'delete', methods: ['DELETE'], requirements: ['id' => Requirement::DIGITS])]
-    public function delete(Emission $emission, EntityManagerInterface $em): Response
+    public function delete(Request $request, Emission $emission, EntityManagerInterface $em): Response
     {
+        $token = $request->request->get('_token');
+        if (!$this->isCsrfTokenValid('delete' . $emission->getId(), $token)) {
+            $this->addFlash('error', 'Jeton CSRF invalide.');
+            return $this->redirectToRoute('admin.emission.index');
+        }
+
         $em->remove($emission);
         $em->flush();
 
         $this->addFlash('success', 'L\'émission a bien été supprimée.');
+
+        // Retourne à l'URL courante (fournie en paramètre)
+        $returnTo = $request->query->get('returnTo');
+        if ($returnTo) {
+            return $this->redirect($returnTo);
+        }
 
         return $this->redirectToRoute('admin.emission.index');
     }
